adds nodes, prints list and clears the space

+++ push_swap/utils/ft_atoi.c
@@ -6,17 +6,17 @@
 /*   By: jabecass <jabecass@student.42lisboa.com    +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2023/01/16 01:23:16 by jabecass          #+#    #+#             */
-/*   Updated: 2023/01/16 01:23:39 by jabecass         ###   ########.fr       */
+/*   Updated: 2023/01/18 14:01:30 by jabecass         ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
 #include "../push_swap.h"
 
-int	ft_atoi(const char *str)
+long int	ft_atoi(const char *str)
 {
-	size_t	i;
-	size_t	sign;
-	size_t	res;
+	long int	i;
+	long int	sign;
+	long int	res;
 
 	i = 0;
 	sign = 1;
